docker-compose/eval-canary/requirements.txt
opensearch-py>=2.0
opensearch-genai-observability-sdk-py[opensearch] @ git+https://github.com/kylehounslow/genai-observability-sdk-py.git@feat/retrieval
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
